Use registered/labelled/named gasses only.
This achieves two things:
1. It adds safety in that you can add your gasses only once and be assured
no random numbers are accidentally used in your dive plan.
2. It allows callers to store a set of gasses and use them repeatedly.
3. It allows the decompressor to specifically know when deco gasses
are available, and use THOSE for accelerated decompression.

The only special rule of "deco gas" is that it cannot be used for
the addDepthChange or addFlat functions. It is exclusively
used by the decompressor to select the best time to switch to them to maximize
decompression safety.

Not that nothing prevents the caller from adding a gas into the
bottom mix list, as well as the deco gas list simultaneously.

4. Finally, I also made a change to store "stages" or "segments"
of the dive as they are added. This is useful for various features
such as knowing the last depth the diver would be at, according to plan.

This allows using that depth as a default for decompression, if one
is not explicitly provided.