Reverting browserify crap. Sorry about this.
This repo is a clean node.js module and
principles of cleanliness tell us not to
mix too many things.

I'm going to create a web-based console
to use this library in REPL mode, but I'll create it
as a separate repo and do it right.

Revert "Proper usage of browserify"

This reverts commit a3e2afdf5eec54f8487d4d767b764f37c657ed11.

Revert "Added an extra make step to browserify the javascript"

This reverts commit 8670283f8e142f4da49df5b2526e7c4082c171a6.